requirements.txt
fpdf2==2.7.8
pandas==3.0.2
PyPDF2==3.0.1
pyshark==0.6
pywin32
requests==2.33.1
rich==14.3.3